+++ flex/index.ts
@@ -5,4 +5,5 @@ export * from "./token";
 export * from "./trader";
 export * from "./config";
 export * from "./web3";
+export * from "./error";
 export * from "../contracts/account-ex";