+++ flex/trader/cancel-order.ts
@@ -1,19 +1,23 @@
 import {
     FlexClientAccount,
     FlexWalletAccount,
+    PRICE_XCHG_ERROR,
     PriceXchgAccount,
     resolveContractError,
     XchgPairAccount,
+    AccountClass,
 } from "../../contracts";
 import { getWallet } from "./internals";
 import { PriceXchgGetDetailsOutput } from "../../contracts/generated/PriceXchgAccount";
-import { TraderOptions } from "./types";
+import { PriceOrder, TraderOptions } from "./types";
 import { Evr, TokenValue } from "../web3";
 import { abiContract, ProcessingErrorCode, TvmErrorCode } from "@eversdk/core";
-import { AccountClass } from "../../contracts/account-ex";
 import { resolveDerivativeTransaction, SdkError } from "./processing";
 import { DerivativeTransactionMessage } from "../web3/accounts";
 import { priceToUnits } from "../web3/utils";
+import { Flex } from "../flex";
+import { queryOrders } from "./query";
+import { FlexError } from "../error";
 
 export type CancelOrderOptions = {
     /**
@@ -124,6 +128,21 @@ function cancelOrderError(error: Error): CancelOrderResult {
     };
 }
 
+export type CancelAllOrdersParams = {
+    /**
+     * Flex Client address. If you are a trader, ask the person who lent you the money.
+     */
+    clientAddress: string;
+    /**
+     * Trader info
+     */
+    trader: TraderOptions;
+};
+
+export type CancelAllOrdersResult = {
+    orders: PriceOrder[];
+};
+
 export async function cancelOrder(
     evr: Evr,
     options: CancelOrderOptions,
@@ -149,7 +168,10 @@ export async function cancelOrder(
     } else if (findOrder(options.orderId, priceDetails.buys)) {
         sell = false;
     } else {
-        throw new Error(`Order ${options.orderId} not found in price ${priceDetails.address}.`);
+        throw new FlexError(
+            PRICE_XCHG_ERROR.canceled.exitCode,
+            `Order ${options.orderId} not found in price ${priceDetails.address}.`,
+        );
     }
     const wallet = await getWallet(evr, {
         marketAddress: options.marketAddress,
@@ -233,7 +255,8 @@ async function getPriceDetails(
         })
     ).output.value0;
     if (!(await evr.accounts.isActive(address))) {
-        throw new Error(
+        throw new FlexError(
+            PRICE_XCHG_ERROR.incorrect_price.exitCode,
             `Orderbook's price account [${address}] does not exist. Please check that the price (${JSON.stringify(
                 price,
             )}) is correct.`,
@@ -407,5 +430,34 @@ function resolveStartingError(original: SdkError, result: CancelOrderResult): Ca
     return {
         status: CancelOrderStatus.ERROR,
         error: error,
+    };
+}
+
+export async function cancelAllOrders(
+    flex: Flex,
+    options: CancelAllOrdersParams,
+): Promise<CancelAllOrdersResult> {
+    const existing = await queryOrders(flex, options.trader.id);
+    const orders: PriceOrder[] = [];
+    for (const order of existing) {
+        const result = await cancelOrder(flex.evr, {
+            marketAddress: order.pair.address,
+            clientAddress: options.clientAddress,
+            trader: options.trader,
+            orderId: order.orderId,
+            price: order.price,
+            waitForOrderbookUpdate: true,
+        });
+        if (result.status === CancelOrderStatus.ERROR) {
+            throw result.error;
+        }
+        orders.push({
+            pairAddress: order.pair.address,
+            price: order.price.toString(),
+            orderId: order.orderId,
+        });
+    }
+    return {
+        orders,
     };
 }