+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 
 
+## [0.10.0] – 2022-11-23
+
+### New
+
+- Integration tests.
+- `Trader.cancelAllOrders` - cancels all orders for the specified trader.
+
+
 ## [0.9.0] – 2022-11-23
 
 ### New